requirements.txt
Django
# django-crispy-forms
# pytz
psycopg2-binary
psycopg2
django-environ
django-allauth==0.44.0